+++ testutil/Fixtures/DummyServiceC.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace TestUtil\Fixtures;
+
+readonly class DummyServiceC
+{
+}