+++ testutil/Fixtures/DummyServiceD.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TestUtil\Fixtures;
+
+readonly class DummyServiceD
+{
+    protected function __construct()
+    {
+    }
+}