+++ testutil/Fixtures/DummyServiceE.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TestUtil\Fixtures;
+
+abstract readonly class DummyServiceE
+{
+    public function __construct()
+    {
+    }
+}