+++ testutil/Fixtures/DummyServiceF.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TestUtil\Fixtures;
+
+abstract readonly class DummyServiceF
+{
+    private function __construct()
+    {
+    }
+}